DRAKEN-4673: Use dept44 hash-utils instead of custom util.
