+++ pom.xml
@@ -4,7 +4,7 @@
 	<parent>
 		<groupId>se.sundsvall.dept44</groupId>
 		<artifactId>dept44-service-parent</artifactId>
-		<version>8.0.8</version>
+		<version>8.0.9</version>
 		<relativePath />
 		<!-- lookup parent from the repository -->
 	</parent>
