+++ src/test/java/se/sundsvall/casedata/integration/db/AttachmentHashConsistencyTest.java
@@ -3,8 +3,6 @@ package se.sundsvall.casedata.integration.db;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import java.nio.charset.StandardCharsets;
-import java.security.MessageDigest;
-import java.util.HexFormat;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -13,6 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import se.sundsvall.casedata.Application;
 import se.sundsvall.casedata.integration.db.model.AttachmentEntity;
 import se.sundsvall.casedata.service.util.BlobBuilder;
+import se.sundsvall.dept44.util.HashUtils;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.MOCK;
@@ -20,8 +19,8 @@ import static se.sundsvall.casedata.TestUtil.MUNICIPALITY_ID;
 import static se.sundsvall.casedata.TestUtil.NAMESPACE;
 
 /**
- * Verifies that the SHA-256 content hash computed in the application (JVM {@code MessageDigest}, lower-case hex) is
- * identical to the hash the database computes over the same stored bytes ({@code MariaDB SHA2(content, 256)}).
+ * Verifies that the SHA-256 content hash computed in the application ({@link HashUtils}, lower-case hex) is identical
+ * to the hash the database computes over the same stored bytes ({@code MariaDB SHA2(content, 256)}).
  *
  * <p>
  * This guards the migration assumption (DRAKEN-4446) that a database-side backfill of the {@code hash} column yields
@@ -42,7 +41,7 @@ class AttachmentHashConsistencyTest {
 	private EntityManager entityManager;
 
 	@Test
-	void applicationHashMatchesDatabaseSha2() throws Exception {
+	void applicationHashMatchesDatabaseSha2() {
 		// Arrange - genuinely binary content (UTF-8 text plus embedded NUL/0xFF/control bytes), stored as a blob with the
 		// hash computed exactly as the application does (SHA-256, lower-case hex).
 		final var text = "Bilaga med abc 123 och text".getBytes(StandardCharsets.UTF_8);
@@ -52,7 +51,7 @@ class AttachmentHashConsistencyTest {
 		content[text.length + 1] = 0x01;
 		content[text.length + 2] = (byte) 0xFF;
 		content[text.length + 3] = 0x02;
-		final var applicationHash = HexFormat.of().formatHex(MessageDigest.getInstance("SHA-256").digest(content));
+		final var applicationHash = HashUtils.sha256Hex(content);
 
 		final var entity = AttachmentEntity.builder()
 			.withMunicipalityId(MUNICIPALITY_ID)
